split dataclass
