+++ rosbag2_pytorch_data_loader/rosbag2_pytorch_dataset.py
@@ -5,35 +5,7 @@ from mcap.reader import NonSeekingReader
 from yaml import safe_load  # type: ignore
 from rosbag2_pytorch_data_loader.exception import TaskDescriptionError
 from rosbag2_pytorch_data_loader.conversion import decode_image_message
-
-from dataclasses import dataclass, field
-from dataclass_wizard import YAMLWizard
-
-
-@dataclass
-class ImageTopicConfig(YAMLWizard):  # type: ignore
-    topic_name: str = ""
-    compressed: bool = True
-
-
-@dataclass
-class ImageOnlyConfig(YAMLWizard):  # type: ignore
-    dataset_type: str = "image_only"
-    image_topics: list[ImageTopicConfig] = field(default_factory=list)
-
-    def get_image_topics(self) -> list[str]:
-        topics: list[str] = []
-        for topic in self.image_topics:
-            topics.append(topic.topic_name)
-        return topics
-
-    def compressed(self, name: str) -> bool:
-        for topic in self.image_topics:
-            if topic.topic_name == name:
-                return topic.compressed
-        raise TaskDescriptionError(
-            "Topic : " + name + " does not exist in task description"
-        )
+from rosbag2_pytorch_data_loader.task_description import ImageOnlyConfig
 
 
 class Rosbag2Dataset(Dataset):  # type: ignore
